# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Google Sheets Proxy API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://google-sheets-d8kr9e2zp-anthony-swifts-projects.vercel.app"
    }
  ],
  "paths": {
    "/read": {
      "post": {
        "summary": "Read data from a Google Sheet",
        "operationId": "readData",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" },
                  "range": { "type": "string" }
                },
                "required": ["spreadsheetId", "range"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Data read from the sheet" }
        }
      }
    },
    "/write": {
      "post": {
        "summary": "Write data to a Google Sheet",
        "operationId": "writeData",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" },
                  "range": { "type": "string" },
                  "values": {
                    "type": "array",
                    "items": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                },
                "required": ["spreadsheetId", "range", "values"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Data written successfully" }
        }
      }
    },
    "/format": {
      "post": {
        "summary": "Format cells in a Google Sheet",
        "operationId": "formatCells",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": {
                    "type": "string",
                    "description": "The ID of the spreadsheet"
                  },
                  "sheetId": {
                    "type": "integer",
                    "description": "The numeric ID of the target sheet (not the title)"
                  },
                  "formatRequests": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "repeatCell": {
                          "type": "object",
                          "properties": {
                            "range": {
                              "type": "object",
                              "properties": {
                                "sheetId": { "type": "integer" },
                                "startRowIndex": { "type": "integer" },
                                "endRowIndex": { "type": "integer" },
                                "startColumnIndex": { "type": "integer" },
                                "endColumnIndex": { "type": "integer" }
                              },
                              "required": ["sheetId", "startRowIndex", "endRowIndex"]
                            },
                            "cell": {
                              "type": "object",
                              "properties": {
                                "userEnteredFormat": {
                                  "type": "object",
                                  "properties": {
                                    "textFormat": {
                                      "type": "object",
                                      "properties": {
                                        "bold": { "type": "boolean" }
                                      },
                                      "required": ["bold"]
                                    }
                                  },
                                  "required": ["textFormat"]
                                }
                              },
                              "required": ["userEnteredFormat"]
                            },
                            "fields": { "type": "string" }
                          },
                          "required": ["range", "cell", "fields"]
                        }
                      },
                      "required": ["repeatCell"]
                    }
                  }
                },
                "required": ["spreadsheetId", "sheetId", "formatRequests"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Cells formatted successfully" }
        }
      }
    },
    "/createSheet": {
      "post": {
        "summary": "Create a new sheet",
        "operationId": "createSheet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" },
                  "title": { "type": "string" }
                },
                "required": ["spreadsheetId", "title"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Sheet created successfully" }
        }
      }
    },
    "/renameSheet": {
      "post": {
        "summary": "Rename a sheet",
        "operationId": "renameSheet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" },
                  "sheetId": { "type": "integer" },
                  "newTitle": { "type": "string" }
                },
                "required": ["spreadsheetId", "sheetId", "newTitle"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Sheet renamed successfully" }
        }
      }
    },
    "/getSheets": {
      "get": {
        "summary": "Get list of sheets",
        "operationId": "getSheets",
        "parameters": [
          {
            "name": "spreadsheetId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Sheet list returned" }
        }
      }
    },
    "/addChart": {
      "post": {
        "summary": "Add a chart to a sheet",
        "operationId": "addChart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" },
                  "sheetId": { "type": "integer" },
                  "chartSpec": { "type": "object" }
                },
                "required": ["spreadsheetId", "sheetId", "chartSpec"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Chart added successfully" }
        }
      }
    },
    "/addPivotTable": {
      "post": {
        "summary": "Add a pivot table",
        "operationId": "addPivotTable",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" },
                  "sheetId": { "type": "integer" },
                  "pivotTableSpec": { "type": "object" },
                  "destination": { "type": "object" }
                },
                "required": ["spreadsheetId", "sheetId", "pivotTableSpec"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Pivot table added successfully" }
        }
      }
    },
    "/getDateTime": {
      "get": {
        "summary": "Get current UTC and local time",
        "operationId": "getDateTime",
        "responses": {
          "200": { "description": "Returns current datetime" }
        }
      }
    },
    "/refreshDirectory": {
      "post": {
        "summary": "Refresh the Directory sheet",
        "operationId": "refreshDirectory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spreadsheetId": { "type": "string" }
                },
                "required": ["spreadsheetId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Directory refreshed" }
        }
      }
    }
  }
}